+++ analysis/Event.cc
@@ -21,10 +21,10 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "BDSOutputROOTEventCollimator.hh"
 #include "BDSOutputROOTEventCoords.hh"
-#include "BDSOutputROOTEventExit.hh"
 #include "BDSOutputROOTEventHistograms.hh"
 #include "BDSOutputROOTEventInfo.hh"
 #include "BDSOutputROOTEventLoss.hh"
+#include "BDSOutputROOTEventLossWorld.hh"
 #include "BDSOutputROOTEventTrajectory.hh"
 #include "BDSOutputROOTEventSampler.hh"
 
@@ -86,8 +86,8 @@ void Event::CommonCtor()
   Eloss           = new BDSOutputROOTEventLoss();
   ElossVacuum     = new BDSOutputROOTEventLoss();
   ElossTunnel     = new BDSOutputROOTEventLoss();
-  ElossWorld      = new BDSOutputROOTEventLoss();
-  ElossWorldExit  = new BDSOutputROOTEventExit();
+  ElossWorld      = new BDSOutputROOTEventLossWorld();
+  ElossWorldExit  = new BDSOutputROOTEventLossWorld();
   PrimaryFirstHit = new BDSOutputROOTEventLoss();
   PrimaryLastHit  = new BDSOutputROOTEventLoss();
   TunnelHit       = new BDSOutputROOTEventLoss();