+++ include/BDSOutputROOTEventLossWorld.hh
@@ -16,11 +16,11 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 */
-#ifndef BDSOUTPUTROOTEVENTEXIT_H
-#define BDSOUTPUTROOTEVENTEXIT_H
+#ifndef BDSOUTPUTROOTEVENTLOSSWORLD_H
+#define BDSOUTPUTROOTEVENTLOSSWORLD_H
 
 #ifndef __ROOTBUILD__
-class BDSHitVolumeExit;
+class BDSHitEnergyDepositionGlobal;
 #endif
 
 #include "TObject.h"
@@ -28,39 +28,44 @@ class BDSHitVolumeExit;
 #include <vector>
 
 /**
- * @brief Data stored for volume exit hits per event.
+ * @brief Data stored for world hits per event.
+ *
+ * These are for loss hits that have no concept of S or curvilinear
+ * coordinates.
  * 
  * @author Laurie Nevay
  */
 
-class BDSOutputROOTEventExit: public TObject
+class BDSOutputROOTEventLossWorld: public TObject
 {
 public:
-  int                n = 0;   ///< Number of entries
-  std::vector<float> totalEnergy;  ///< Energy deposited in step
+  int                n = 0;                 ///< Number of entries.
+  std::vector<float> totalEnergy;           ///< Energy deposited in step.
+  std::vector<float> preStepKineticEnergy;  ///< Kinetic energy in GeV at pre step point.
   std::vector<float> postStepKineticEnergy; ///< Kinetic energy in GeV at post step point.
+  std::vector<float> stepLength;            ///< Step length.
 
   //// @{ Global coordinates
   std::vector<float> X;
   std::vector<float> Y;
   std::vector<float> Z;
   /// @}
-  std::vector<float> T; ///< Global time (time since beginning of event).
-
+  
+  std::vector<float> T;       ///< Global time (time since beginning of event).
   std::vector<int>   partID;  ///< ParticleID that create the deposit
   std::vector<int>   trackID; ///< TrackID that created the deposit
   std::vector<int>   parentID;///< ParentID that created the deposit
   std::vector<float> weight;  ///< Weight
   std::vector<int>   turn;    ///< Turn number
   
-  BDSOutputROOTEventExit();
-  virtual ~BDSOutputROOTEventExit();
+  BDSOutputROOTEventLossWorld();
+  virtual ~BDSOutputROOTEventLossWorld();
 #ifndef __ROOTBUILD__
-  void Fill(const BDSHitVolumeExit* hit);
+  void Fill(const BDSHitEnergyDepositionGlobal* hit);
 #endif
   virtual void Flush();
 
-  ClassDef(BDSOutputROOTEventExit,1);
+  ClassDef(BDSOutputROOTEventLossWorld,1);
 };